backend/requirements.txt
fastapi
uvicorn
httpx
beautifulsoup4
openai
playwright
python-dotenv
pydantic
slowapi
python-multipart
